+++ ChangeLog.md
@@ -1,6 +1,7 @@
 Version 1.4.0
 -------------
 * Now requires Visual C++ Redistributable for Visual Studio 2017
+* Added support for GetRootAndLogicalPrefixPaths from IVssBackupComponentsEx4.
 * Distributed as a NuGet package.
 * Added a build targetting .NET 4.0
 * Internal reorganization of build process.